# stats: refactoring MetricImpl without strings