Signed-off-by: James Buckland <jbuckland@google.com>

*Description*: This reworks `MetricImpl` to get rid of the `std::string` it keeps as a private member variable. Since `MetricImpl` never gets instantiated as-is, I let each of its extending classes implement `name()` for themselves, often without the need to copy strings.  (Part of https://github.com/envoyproxy/envoy/issues/3585/)

In my estimation, running 

    bazel build //source/exe:envoy-static --copt='-ltcmalloc'
    timeout 90 bazel-bin/source/exe/envoy-static --disable-hot-restart --config-path <a_config_with_10k_clusters>
    pprof -text bazel-bin/source/exe/envoy-static main_common_base*

on upstream as-is vs. this branch saves ~20% memory consumption on the heap. 

*Risk Level*: Medium -- anything that changes stats is tricky. But this doesn't change RawStatData, just `MetricImpl`.
*Testing*: This required editing the `MockCounter`/`MockGauge` classes, but otherwise it passes all tests.
*Docs Changes*: N/A
*Release Notes*: N/A