@ambuc 2018-08-17
OK, here are some more performance testing results:

(all run under `--disable-hot-restart`)

10^ | Clusters | Run for | Upstream | This Branch | Reduction
------------ | ------------ | ------------ | ------------- | ------------- | -------------
1 | 10 | 30 sec | 2534.52kB | 2480.34kB | 97.9%
2 | 100 | 60 sec | 8530.54kB | 7990.80kB | 93.7%
3 | 1000  | 90 sec | 68589.73kB | 63132.50kB | 92.0%
4 | 10000 | 90 sec | 4366736.40kB | 3566095.10 kB | 81.7%