@jmarantz 2018-08-17
cool, although I think the reduction would be 100-97.9%, 100-93.7%, etc. Still awesome, as this reduces by 18.3% the scaled case.