@jmarantz 2018-08-17
one more thing while we wait for a maintainer to review: you measured memory high-water here, right?  Do you want to leave some notes somewhere (in the PR and maybe also in some developer-oriented doc) about how you did that so they can be repro'd by others?

Also, curious what impact this has on CPU and elapsed time for these tests.