@ambuc 2018-08-17
Sounds good to me -- I will do that in a separate PR, so that we can figure out where in the repo to put it (at our leisure).