@dnoe 2018-08-17
The PR description says that `MetricImpl` is never instantiated directly. But I don't see any comments indicating that it is never intended to be instantiated directly. I think the way things are structured right now it has the pure virtual name() so at least it shouldn't compile rather than failing mysteriously at runtime. Do you want to add a comment to the MetricImpl header file too?