@ambuc 2018-08-17
@dnoe Sounds good, I'll put a comment to that effect -- this might happen to other string private member variables, but it does make sense to have a non-virtual `MetricImpl` for encapsulating the `Flags` logic, at least.