@PiotrSikora 2018-08-18
@ambuc I'm a bit confused by the numbers presented in https://github.com/envoyproxy/envoy/pull/4190#issuecomment-413884648, how come memory usage doesn't grow proportionally to number of clusters?

For the record, I see decent improvements with this change (~8%), but in my tests the allocated memory is ~72kB per static cluster, and I'm getting ~73MB for 1000 clusters and ~717MB for 10000 clusters...

How come you're getting 5x more for 10000 clusters? Where are you getting the memory usage from?