@ambuc 2018-08-20
@PiotrSikora I don't know why the nonlinearity, but here's an experiment I just ran:

I have two cluster configs with 10k clusters. In the smaller, each cluster name is ~10 characters long; in the larger, each cluster name is ~50 characters long. 

| clusters            | 10k     | 10k     |
|---------------------|---------|---------|
| average cluster name length | ~10     | ~50     |
| allocated           | 3.56 GB | 5.51 GB |  

This difference is too large -- maybe this explains the 5x increase we saw above? There's something about string allocations I don't understand which _may_ be driving this nonlinearity.