@PiotrSikora 2018-08-20
@ambuc in my tests the cluster name is ~18 characters long and the memory usage is ~717MB (~72kB per cluster), so that's not it.

There is also per thread/worker overhead of about 1850 bytes, but for the math to work for that, you'd need to be running your tests with `--concurrency 160` or similar.